+++ test/pagy/i18n_locales_test.rb
@@ -21,10 +21,11 @@ describe 'pagy/locales' do
   Pagy.root.join('locales').each_child do |f|
     next unless f.extname == '.yml'
 
-    message = "locale file #{f}"
-    locale  = f.basename.to_s[0..-5]
-    comment = f.readlines.first.to_s.strip
-    rule    = comment.to_s.split[1][1..].to_s.to_sym
+    message       = "locale file #{f}"
+    locale        = f.basename.to_s[0..-5]                 # e.g. de
+    comment       = f.readlines.first.to_s.strip           # e.g. :one_other pluralization (see https://github.com/ddnexus/pagy/blob/master/gem/lib/pagy/i18n.rb)
+    rule          = comment.to_s.split[1][1..].to_s.to_sym # e.g. one_other
+    language_yml  = YAML.safe_load(f.read)
 
     it 'includes a comment with the pluralization rule and the i18n.rb reference' do
       _(rules).must_include rule, message
@@ -34,8 +35,7 @@ describe 'pagy/locales' do
       _(Pagy::I18n::P11n::LOCALE[locale]).must_equal Pagy::I18n::P11n::RULE[rule], message
     end
     it 'pluralizes item_name according to the rule' do
-      hash      = YAML.safe_load(f.read)
-      item_name = hash[locale]['pagy']['item_name']
+      item_name = language_yml[locale]['pagy']['item_name']
       case item_name
       when String
         _(rule).must_equal :other
@@ -43,6 +43,22 @@ describe 'pagy/locales' do
         _(item_name.keys - counts[rule]).must_be_empty
       else
         raise StandardError, "item_name must be Hash or String"
+      end
+    end
+    it "ensures #{locale}.yml has the correct aria_label,nav and item_name keys per the declared (#{rule}) rule" do
+      skip if %w[ta sw].include?(locale) # ta.yml and sw.yml do not have the requisite keys yet
+
+      pluralizations = counts[rule]
+
+      if rule == :other
+        # For the :other rules, we should not have any keys under the
+        # ['pagy']['item_name'] and ['pagy']['aria_label']['nav'] hierarchies.
+        # We should just have a String.
+        _(language_yml[locale]['pagy']['item_name']).must_be_instance_of(String)
+        _(language_yml[locale]['pagy']['aria_label']['nav']).must_be_instance_of(String)
+      else
+        _(language_yml[locale]['pagy']['item_name'].keys.sort).must_equal pluralizations.sort, "In #{message} - check that ['pagy']['item_name'] does not have keys inconsistent with #{rule}"
+        _(language_yml[locale]['pagy']['aria_label']['nav'].keys.sort).must_equal pluralizations.sort, "In #{message} - check that ['pagy']['aria_label']['nav'] does not have keys inconsistent with #{rule}"
       end
     end
   end